+++ ui/src/ModalFormComponents/ColorCircle.tsx
@@ -16,7 +16,7 @@
  */
 
 import React from 'react';
-import {Color} from '../Roles/Role.interface';
+import {Color} from '../Roles/RoleTag.interface';
 import './ColorCircle.scss';
 
 interface Props {
