+++ ui/src/Roles/MyRolesForm.tsx
@@ -23,7 +23,7 @@ import {GlobalStateProps} from '../Redux/Reducers';
 import {setAllGroupedTagFilterOptionsAction} from '../Redux/Actions';
 import {FilterOption} from '../CommonTypes/Option';
 import {AllGroupedTagFilterOptions} from '../ReusableComponents/ProductFilter';
-import {Color, RoleTag} from './Role.interface';
+import {Color, RoleTag} from './RoleTag.interface';
 import RoleClient from './RoleClient';
 import {Tag} from '../Tags/Tag.interface';
 import {Space} from '../Space/Space';
